[refactor] MenuLineItem DTO를 외부 클래스로 분리
Request, Command DTO로 분리하는 과정에서 MenuLineItem 을 외부 클래스로 분리함
